refactor: split furigana-annotator.js into multiple files (≤300 lines each)
Split the 900-line monolithic furigana-annotator.js into 4 focused files:
- furigana-annotator.js (293 lines): main class, core utils, _handleCallNode
- furigana-node-handlers.js (244 lines): variable/literal/control flow handlers
- furigana-call-helpers.js (282 lines): call annotation helpers, extension handlers
- furigana-label-map.js (239 lines): all label maps and static data

Uses Object.assign mixin pattern to compose handler methods onto the prototype.
All 254 existing tests pass with no changes.

Refs #283
